+++ noderedpy/__init__.py
@@ -7,7 +7,7 @@ from ._property import (
 from ._server import Server, StandaloneServer
 
 
-__version__ = "0.1.2"
+__version__ = "0.1.3"
 __author__ = "oyajiDev"
 __email__ = "this.dev.somehit@gmail.com"