+++ noderedpy/templates/__init__.py
@@ -1,24 +1,19 @@
 # -*- coding: utf-8 -*-
-import json, noderedpy
-from ._property import (
+import os, json, noderedpy
+from .._property import (
     InputProperty, ListProperty, DictProperty,
     SpinnerProperty, ComboBoxProperty
 )
+from . import __path__
 
 
 def package_json(node:"noderedpy._nodered.Node") -> str:
-    return json.dumps({
-        "name": node.name,
-        "version": "1.0.0",
-        "description": f"nodered.py {node.name} node",
-        "author": "nodered.py",
-        "keywords": [ "node-red" ],
-        "node-red": {
-            "nodes": {
-                node.name: f"lib/{node.name}.js"
-            }
-        }
-    }, indent = 4)
+    with open(os.path.join(__path__[0], "template.json"), "r", encoding = "utf-8") as tr:
+        tt = tr.read()
+
+    return tt.replace(
+        "{$node_name}", node.name
+    )
 
 def node_html(node:"noderedpy._nodered.Node") -> str:
     properties_html, properties_js, properties_js_prepare, properties_js_cancel, properties_js_save = [], [], [], [], []
@@ -40,7 +35,7 @@ def node_html(node:"noderedpy._nodered.Node") -> str:
         <label><i class="{property.display_icon}"></i> <span>{property.display_name}</span></label>
     </div>
     <div class="form-row node-input-{property.name}-container-row">
-        <ol id="node-input-{property.name}-container" style="height:250px;"></ol>
+        <ol id="node-input-{property.name}-container" style="height:{property.height}px;"></ol>
     </div>
 """)
             properties_js_prepare.append('''
@@ -72,7 +67,7 @@ def node_html(node:"noderedpy._nodered.Node") -> str:
         <label><i class="{property.display_icon}"></i> <span>{property.display_name}</span></label>
     </div>
     <div class="form-row node-text-editor-row">
-        <div style="height:250px;" class="node-text-editor" id="node-input-{property.name}"></div>
+        <div style="height:{property.height}px;" class="node-text-editor" id="node-input-{property.name}"></div>
     </div>
 """)
             properties_js_prepare.append('''
@@ -102,8 +97,18 @@ def node_html(node:"noderedpy._nodered.Node") -> str:
         <input type="text" id="node-input-{property.name}" style="width:calc(100% - 22px);">
     </div>
 """)
+            spinner_configs = []
+            if property.min:
+                spinner_configs.append(f"                min: {property.min},")
+            if property.max:
+                spinner_configs.append(f"                max: {property.max},")
+            if property.step:
+                spinner_configs.append(f"                step: {property.step}")
+
             properties_js_prepare.append('''
-            $("#node-input-''' + property.name + '''").spinner({ min: 0 });
+            $("#node-input-''' + property.name + '''").spinner({
+                ''' + "\n".join(spinner_configs) + '''
+            });
 ''')
             default_value = str(property.default)
         elif isinstance(property, ComboBoxProperty):
@@ -135,130 +140,33 @@ def node_html(node:"noderedpy._nodered.Node") -> str:
         if default_value:
             properties_js.append("            " + property.name + ': { value: ' + default_value + ' }')
 
-    return '''
-<script type="text/html" data-template-name="''' + node.name + '''">
-    <style>
-        span.ui-spinner {
-            width: 100%;
-        }
-    </style>
-    <div class="form-row">
-        <label for="node-input-name"><i class="fa fa-tag"></i> <span>Name</span></label>
-        <input type="text" id="node-input-name" style="width:calc(100% - 105px);">
-    </div>
-    <hr>
-    ''' + "".join(properties_html) + '''
-</script>
+    with open(os.path.join(__path__[0], "template.html"), "r", encoding = "utf-8") as tr:
+        tt = tr.read()
 
-<script type="text/javascript">
-    RED.nodes.registerType("''' + node.name + '''", {
-        category: "''' + node.category + '''",
-        color: "#FDD0A2",
-        defaults: {
-            name: { value: "" },
-''' + ",\n".join(properties_js) + '''
-        },
-        inputs: 1, outputs: 1,
-        icon: "function.png",
-        label: function() {
-            return this.name;
-        },
-        oneditprepare: function() {
-''' + "\n".join(properties_js_prepare) + '''
-        },
-        oneditcancel: function() {
-''' + "\n".join(properties_js_cancel) + '''
-        },
-        oneditsave: function() {
-''' + "\n".join(properties_js_save) + '''
-        }
-    });
-</script>
-'''
+    return tt.replace(
+        "{$node_name}", node.name
+    ).replace(
+        "{$node_category}", node.category
+    ).replace(
+        "{$properties_html}", "".join(properties_html)
+    ).replace(
+        "{$properties_js}", ",\n".join(properties_js)
+    ).replace(
+        "{$properties_js_prepare}", "\n".join(properties_js_prepare)
+    ).replace(
+        "{$properties_js_cancel}", "\n".join(properties_js_cancel)
+    ).replace(
+        "{$properties_js_save}", "\n".join(properties_js_save)
+    )
 
 def node_js(node:"noderedpy._nodered.Node", port:int) -> str:
-    return '''
-let messageCache = {};
-    
-module.exports = function(RED) {
-    function fnNode(config) {
-        var node = this;
-        RED.nodes.createNode(this, config);
+    with open(os.path.join(__path__[0], "template.js"), "r", encoding = "utf-8") as tr:
+        tt = tr.read()
 
-        this.status({ fill: "blue", shape: "dot", text: "Ready" });
-        this.on("input", (message) => {
-            messageCache._msgid = message._msgid;
-            delete message._msgid;
-
-            var reqToSend = null;
-            if (message.req != undefined && typeof(message.req) == "object") {
-                messageCache.req = message.req;
-
-                reqToSend = {
-                    payload: message.req.payload,
-                    body: message.req.body,
-                    cookie: message.req.cookie,
-                    header: {}
-                };
-                for (var idx = 0; idx < message.req.rawHeaders.length / 2; idx++) {
-                    reqToSend.header[message.req.rawHeaders[idx * 2]] = message.req.rawHeaders[idx * 2 + 1];
-                }
-
-                message.req = reqToSend;
-            }
-            if (message.res != undefined && typeof(message.res) == "object") {
-                messageCache.res = message.res;
-                delete message.res;            
-            }
-
-            var configToSend = {};
-            for (var name of ''' + str([ property.name for property in node.properties]) + ''') {
-                var config_item = config[name];
-                if (typeof(config_item) == "string" && (config_item.startsWith("{") && config_item.endsWith("}"))) {
-                    config_item = JSON.parse(config_item);
-                }
-                configToSend[name] = config_item;
-            }
-
-            node.status({ fill: "green", shape: "dot", text: "Running" });
-
-            try {
-                fetch("http://127.0.0.1:''' + str(port) + '''/nodes/''' + node.name + '''", {
-                    method: "POST",
-                    headers: { "Content-Type": "application/json" },
-                    body: JSON.stringify({
-                        props: configToSend, msg: message
-                    })
-                }).then(async (resp) => {
-                    var message = await resp.json();
-                    const state = message.state;
-                    delete message.state;
-
-                    if (state == "success") {
-                        message._msgid = messageCache._msgid;
-                        if (messageCache.req != undefined) {
-                            message.req = messageCache.req;
-                        }
-                        if (messageCache.res != undefined) {
-                            message.res = messageCache.res;
-                        }
-
-                        node.send(message);
-                        node.status({ fill: "green", shape: "dot", text: "Finished" });
-                    }
-                    else {
-                        node.error(message.message);
-                        node.status({ fill: "red", shape: "dot", text: "Stopped, see debug panel" });
-                    }
-                });
-            }
-            catch (err) {
-                node.error(err.message);
-                node.status({ fill: "red", shape: "dot", text: "Stopped, see debug panel" });
-            }
-        });
-    }
-
-    RED.nodes.registerType("''' + node.name + '''", fnNode);
-}
-'''
+    return tt.replace(
+        "{$node_name}", node.name
+    ).replace(
+        "{$node_properties}", str([ property.name for property in node.properties])
+    ).replace(
+        "{$port}", str(port)
+    )